requirements.txt
openai 
torch --index-url https://download.pytorch.org/whl/cpu
sentence-transformers 
faiss-cpu 
rank_bm25
hf_xet 
sentencepiece
pymupdf
bert-score
scikit-learn
nltk
evaluate
transformers
llama-cpp-python